refactor(backend): rename users handler_utils module
