+++ backend/src/users/service/handler_utils.py
